# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Availability:
    properties:
      capacity:
        type: integer
      date:
        type: string
      id:
        type: string
      reserved:
        type: integer
      restaurant_id:
        type: string
      time_slot:
        type: string
      updated_at:
        type: string
    type: object
  domain.Booking:
    properties:
      alternatives:
        items:
          $ref: '#/definitions/domain.BookingAlternative'
        type: array
      comment:
        type: string
      completed_at:
        type: string
      confirmed_at:
        type: string
      created_at:
        type: string
      date:
        type: string
      duration:
        type: integer
      guests_count:
        type: integer
      id:
        type: string
      rejected_at:
        type: string
      restaurant_id:
        type: string
      status:
        $ref: '#/definitions/domain.BookingStatus'
      time:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  domain.BookingAlternative:
    properties:
      accepted_at:
        type: string
      booking_id:
        type: string
      created_at:
        type: string
      date:
        type: string
      id:
        type: string
      message:
        type: string
      rejected_at:
        type: string
      time:
        type: string
    type: object
...
  domain.Fact:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      restaurant_id:
        type: string
    type: object
...
  domain.Restaurant:
    properties:
      address:
        type: string
      contact_email:
        type: string
      contact_phone:
        type: string
      created_at:
        type: string
      cuisine:
        type: string
      description:
        type: string
      facts:
        items:
          $ref: '#/definitions/domain.Fact'
        type: array
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  handlers.CreateRestaurantRequest:
    properties:
      address:
        type: string
      contact_email:
        type: string
      contact_phone:
        type: string
      cuisine:
        type: string
      description:
        type: string
      facts:
        items:
          type: string
        type: array
      name:
        type: string
    required:
    - address
    - contact_email
    - contact_phone
    - cuisine
    - name
    type: object
...
  handlers.SetAvailabilityRequest:
    properties:
      capacity:
        minimum: 1
        type: integer
      date:
        type: string
      time_slot:
        type: string
    required:
    - capacity
    - date
    - time_slot
    type: object
...
  handlers.UpdateRestaurantRequest:
    properties:
      address:
        type: string
      contact_email:
        type: string
      contact_phone:
        type: string
      cuisine:
        type: string
      description:
        type: string
      name:
        type: string
    required:
    - address
    - contact_email
    - contact_phone
    - cuisine
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@restaurant-booking.com
    name: API Support
  description: API для системы бронирования ресторанов
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Restaurant Booking API
  version: "1.0"
paths:
...
  /restaurants:
    get:
      consumes:
      - application/json
      description: Get a list of all restaurants with optional pagination
      parameters:
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Restaurant'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List restaurants
      tags:
      - restaurants
    post:
      consumes:
      - application/json
      description: Create a new restaurant
      parameters:
      - description: Restaurant data
        in: body
        name: restaurant
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRestaurantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Restaurant'
        "400":
          description: Invalid data
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create restaurant
      tags:
      - restaurants
  /restaurants/{id}:
    delete:
      consumes:
      - application/json
      description: Delete restaurant by ID
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Restaurant not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete restaurant
      tags:
      - restaurants
    get:
      consumes:
      - application/json
      description: Get detailed information about a restaurant by ID
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Restaurant'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Restaurant not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get restaurant
      tags:
      - restaurants
    put:
      consumes:
      - application/json
      description: Update an existing restaurant
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      - description: Restaurant data
        in: body
        name: restaurant
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateRestaurantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Restaurant'
        "400":
          description: Invalid data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Restaurant not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update restaurant
      tags:
      - restaurants
  /restaurants/{id}/availability:
    get:
      consumes:
      - application/json
      description: Get availability for a restaurant on a specific date
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      - description: Date (YYYY-MM-DD)
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Availability'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Restaurant not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get availability
      tags:
      - restaurants
      - availability
    post:
      consumes:
      - application/json
      description: Set availability for a specific date and time
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      - description: Availability data
        in: body
        name: availability
        required: true
        schema:
          $ref: '#/definitions/handlers.SetAvailabilityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Availability'
        "400":
          description: Invalid data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Restaurant not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Set availability
      tags:
      - restaurants
      - availability
  /restaurants/{id}/bookings:
    get:
      consumes:
      - application/json
      description: Get all bookings for a specific restaurant
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      - description: Booking status (pending,confirmed,rejected,canceled,completed)
        in: query
        name: status
        type: string
      - description: Date (YYYY-MM-DD)
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Booking'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Restaurant not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get restaurant bookings
      tags:
      - restaurants
      - bookings
...
schemes:
- http
- https
swagger: "2.0"